requirements.txt
absl-py==0.7.0
astor==0.7.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
Markdown==3.0.1
nltk==3.4
numpy==1.16.1
protobuf==3.6.1
python-dotenv==0.10.1
-e git+https://github.com/kousuke-takeuchi/python-opensubtitles@a4cdba1ef5c8a16d5e94c1b17f0e209e23717af1#egg=python_opensubtitles
PyYAML==3.13
-e git://github.com/datalogai/recurrentshop.git@3d08db80a154fbcf3730f7c7f35d6f7973b4abbc#egg=recurrentshop
scipy==1.2.0
-e git+https://github.com/farizrahman4u/seq2seq@c020ccfc1fa3a651be272f8b4be48a10f9c3f0fa#egg=seq2seq
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm==4.30.0
twitter==1.18.0
Werkzeug==0.14.1
